fix(picker.actions): ensure the current window is updated after tabdrop (#2326)
## Description

Update the winid (in the `win` variable) after `tab drop`.

## Related Issue(s)

Without this patch, when the user set `confirm` to `{ action =
"confirm", cmd = "tabdrop" }` and jump to a location in a different tab,
since the `win` is still pointing to the original window (tab), there'd
be a `Cursor position outside buffer` error.