fix(picker.actions): `drop` and `tabdrop` should never reload existing buffers (#2368)
## Description

When using `tabdrop` as the `jump` action and jumping to a location in
the current buffer, the existing implementation always triggers a reload
of the current buffer because the `vim.cmd` is unconditional. This
causes unnecessary triggering of `BufRead*` and `LspAttach`
auto-commands, which can then trigger a lot of things that can be
CPU-intensive (linters, LSP, etc.). This PR skips calling the `cmd` when
jumping to the current buffer, and hence avoids the excessive triggering
of those auto-commands.

---------

Co-authored-by: Folke Lemaitre <folke.lemaitre@gmail.com>